metrics-performance move metrics methods to a util module
